+++ code.rkt
@@ -5,6 +5,7 @@
 
 (provide css-code
          c-code
+         cpp-code
          csv-code
          html-code
          js-code
@@ -21,6 +22,7 @@
          yaml-code
          cssblock
          cblock
+         cppblock
          csvblock
          htmlblock
          jsblock
@@ -37,6 +39,7 @@
          yamlblock
          cssblock0
          cblock0
+         cppblock0
          csvblock0
          htmlblock0
          jsblock0
@@ -64,6 +67,7 @@
 
 (define-syntax-rule (css-code . rest) (lang:css-code . rest))
 (define-syntax-rule (c-code . rest) (lang:c-code . rest))
+(define-syntax-rule (cpp-code . rest) (lang:cpp-code . rest))
 (define-syntax-rule (csv-code . rest) (lang:csv-code . rest))
 (define-syntax-rule (html-code . rest) (lang:html-code . rest))
 (define-syntax-rule (js-code . rest) (lang:js-code . rest))
@@ -80,6 +84,7 @@
 (define-syntax-rule (yaml-code . rest) (lang:yaml-code . rest))
 (define-syntax-rule (cssblock . rest) (lang:cssblock . rest))
 (define-syntax-rule (cblock . rest) (lang:cblock . rest))
+(define-syntax-rule (cppblock . rest) (lang:cppblock . rest))
 (define-syntax-rule (csvblock . rest) (lang:csvblock . rest))
 (define-syntax-rule (htmlblock . rest) (lang:htmlblock . rest))
 (define-syntax-rule (jsblock . rest) (lang:jsblock . rest))
@@ -96,6 +101,7 @@
 (define-syntax-rule (yamlblock . rest) (lang:yamlblock . rest))
 (define-syntax-rule (cssblock0 . rest) (lang:cssblock0 . rest))
 (define-syntax-rule (cblock0 . rest) (lang:cblock0 . rest))
+(define-syntax-rule (cppblock0 . rest) (lang:cppblock0 . rest))
 (define-syntax-rule (csvblock0 . rest) (lang:csvblock0 . rest))
 (define-syntax-rule (htmlblock0 . rest) (lang:htmlblock0 . rest))
 (define-syntax-rule (jsblock0 . rest) (lang:jsblock0 . rest))